chore: update viteplus

+++ .templates/scaffold.ts
@@ -31,6 +31,10 @@ const scaffold = (type: string) => {
   const targetDir =
     typeof targetArg === 'string' ? path.resolve(process.cwd(), targetArg) : process.cwd();
 
+  if (fs.existsSync(targetDir)) {
+    throw new Error(`Target directory ${targetDir} already exists.`);
+  }
+
   process.stdout.write(`Scaffolding ${type} template in ${targetDir}...\n`);
 
   fs.cpSync(templateDir, targetDir, {
